# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Default module",
    "description": "",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "key"
    },
    {
      "name": "external"
    },
    {
      "name": "internal"
    }
  ],
  "paths": {
    "/keys": {
      "post": {
        "summary": "create key",
        "deprecated": false,
        "description": "",
        "operationId": "create.key",
        "tags": [
          "key",
          "external"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "key": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                },
                "required": [
                  "key"
                ]
              },
              "examples": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "summary": "list keys",
        "deprecated": false,
        "description": "",
        "operationId": "list.keys",
        "tags": [
          "key",
          "external"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Key"
                  }
                }
              }
            },
            "headers": {}
          },
          "204": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/pick": {
      "get": {
        "summary": "pick key",
        "deprecated": false,
        "description": "",
        "operationId": "pick.key",
        "tags": [
          "key",
          "internal"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/keys/{id}": {
      "delete": {
        "summary": "delete key",
        "deprecated": false,
        "description": "",
        "operationId": "delete.key",
        "tags": [
          "key",
          "external"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
      "Key": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "key"
        ]
      },
      "CreateKeyRequest": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "key"
        ]
      }
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}